Address review comment. Move tests to SegmentReplicationIndexShardTests
