+++ server/src/test/java/org/opensearch/index/shard/IndexShardTests.java
@@ -139,7 +139,6 @@ import org.opensearch.indices.breaker.NoneCircuitBreakerService;
 import org.opensearch.indices.fielddata.cache.IndicesFieldDataCache;
 import org.opensearch.indices.recovery.RecoveryState;
 import org.opensearch.indices.recovery.RecoveryTarget;
-import org.opensearch.indices.replication.checkpoint.ReplicationCheckpoint;
 import org.opensearch.indices.replication.checkpoint.SegmentReplicationCheckpointPublisher;
 import org.opensearch.indices.replication.common.ReplicationLuceneIndex;
 import org.opensearch.indices.replication.common.ReplicationType;
@@ -3499,27 +3498,6 @@ public class IndexShardTests extends IndexShardTestCase {
         assertTrue(stop.compareAndSet(false, true));
         thread.join();
         closeShards(newShard);
-    }
-
-    /**
-     *  Test that latestReplicationCheckpoint returns null only for docrep enabled indices
-     */
-    public void testReplicationCheckpointNullForDocRep() throws IOException {
-        Settings indexSettings = Settings.builder().put(IndexMetadata.SETTING_REPLICATION_TYPE, "DOCUMENT").put(Settings.EMPTY).build();
-        final IndexShard indexShard = newStartedShard(false, indexSettings);
-        assertNull(indexShard.getLatestReplicationCheckpoint());
-        closeShards(indexShard);
-    }
-
-    /**
-     *  Test that latestReplicationCheckpoint returns ReplicationCheckpoint for segrep enabled indices
-     */
-    public void testReplicationCheckpointNotNullForSegReb() throws IOException {
-        Settings indexSettings = Settings.builder().put(IndexMetadata.SETTING_REPLICATION_TYPE, "SEGMENT").put(Settings.EMPTY).build();
-        final IndexShard indexShard = newStartedShard(indexSettings);
-        final ReplicationCheckpoint replicationCheckpoint = indexShard.getLatestReplicationCheckpoint();
-        assertNotNull(replicationCheckpoint);
-        closeShards(indexShard);
     }
 
     /**
